+++ src/merger/in_file/merger_n_files_ge64.cpp
@@ -25,7 +25,7 @@ void merge_n_files_greater_than_64_colors(
     // On ouvre tous les fichiers que l'on doit fusionner
     //
     std::vector<stream_reader*> i_files (file_list.size());
-    for(int i = 0; i < file_list.size(); i += 1)
+    for(size_t i = 0; i < file_list.size(); i += 1)
     {
         stream_reader* f = stream_reader_library::allocate( file_list[i] );
         if( f == NULL )
@@ -41,7 +41,7 @@ void merge_n_files_greater_than_64_colors(
     // On cree les buffers pour tamponner les lectures
     //
     std::vector<uint64_t*> i_buffer(i_files.size());
-    for(int i = 0; i < i_files.size(); i += 1)
+    for(size_t i = 0; i < i_files.size(); i += 1)
         i_buffer[i] = new uint64_t[_iBuff_];
 
     //
@@ -54,7 +54,7 @@ void merge_n_files_greater_than_64_colors(
     // disponible dans chacun des flux
     //
     std::vector<int64_t> nElements(i_files.size());
-    for(int i = 0; i < i_files.size(); i += 1)
+    for(size_t i = 0; i < i_files.size(); i += 1)
         nElements[i] = 0;
 
     //
@@ -62,7 +62,7 @@ void merge_n_files_greater_than_64_colors(
     // dans chacun des flux
     //
     std::vector<int64_t> counter(i_files.size());
-    for(int i = 0; i < i_files.size(); i += 1)
+    for(size_t i = 0; i < i_files.size(); i += 1)
         counter[i] = 0;
 
     //
@@ -70,7 +70,7 @@ void merge_n_files_greater_than_64_colors(
     // associée à chacun des flux
     //
     std::vector<int64_t> color_pos(i_files.size());
-    for(int i = 0; i < i_files.size(); i += 1)
+    for(size_t i = 0; i < i_files.size(); i += 1)
         color_pos[i] = i * iSize;
 
     //
@@ -95,7 +95,7 @@ void merge_n_files_greater_than_64_colors(
         //
         // On verifie sur tous les fichiers ouverts que l'on a des données
         //
-        for(int i = 0; i < i_files.size(); i += 1)
+        for(size_t i = 0; i < i_files.size(); i += 1)
         {
             //
             // Doit'on recharger des données dans le flux ?
@@ -135,7 +135,7 @@ void merge_n_files_greater_than_64_colors(
             //
             curr_index = -1;
             curr_value = 0xFFFFFFFFFFFFFFFF; // la premmière du premier flux
-            for(int i = 0; i < i_files.size(); i += 1)
+            for(size_t i = 0; i < i_files.size(); i += 1)
             {
                 const int       pos = counter [i];
                 const uint64_t* buf = i_buffer[i];
